requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
pydantic
python-dotenv
pymongo
openai
motor
